Put for and while in tofix, fix some typings
