+++ packages/engine/src/lib/dataControls/DropdownControl.ts
@@ -31,6 +31,6 @@ export class DropdownControl extends DataControl {
       },
     }
 
-    super(options)
+    super(options as any)
   }
 }
